refactor: default imports and arquitecture folder
