+++ src/app/[locale]/page.tsx
@@ -1,4 +1,4 @@
-import { ExampleComponent } from '@components/ExampleComponent';
+import { ExampleComponent } from '@/components/ExampleComponent';
 
 export default function Home(): JSX.Element {
   return <ExampleComponent />;
