+++ src/config/SCRegistry.tsx
