+++ src/middleware.ts
@@ -1,5 +1,5 @@
+import i18nConfig from 'i18n.config';
 import createMiddleware from 'next-intl/middleware';
-import i18nConfig from '../i18n.config';
 
 export default createMiddleware({
   // A list of all locales that are supported
