+++ src/store/useBoundStore.ts
@@ -1,5 +1,7 @@
 import { create } from 'zustand';
-import { IBoundStoreProps } from '@services/@types/zustand';
+
+import { IBoundStoreProps } from '@/types/zustand';
+
 import { createThemeSlice } from './themeSlice';
 
 export const useBoundStore = create<IBoundStoreProps>()((...a) => ({
