+++ src/types/jest.d.ts
@@ -1,0 +1,14 @@
+/// <reference types="jest" />
+
+declare namespace jest {
+  interface It {
+    customRender: <
+      Q extends Queries = typeof queries,
+      Container extends Element | DocumentFragment = HTMLElement,
+      BaseElement extends Element | DocumentFragment = Container
+    >(
+      ui: React.ReactElement,
+      options?: RenderOptions<Q, Container, BaseElement>
+    ) => RenderResult<Q, Container, BaseElement>;
+  }
+}
